+++ Internal peripherals/ADC/PIC16F877A_ADC.h
@@ -7,7 +7,7 @@
 #define _XTAL_FREQ 16000000
 
 void ADC_Init(uint8_t* adc_channel, uint8_t len);
-uint8_t ADC_Read(uint8_t adc_channel);
+uint16_t ADC_Read(uint8_t adc_channel);
 float ADC_lm35toDeg(uint16_t negative_deg, uint16_t positive_deg);
 
 #endif
